BUG: Fix timestamp type/format mismatch

+++ core/comm.c
@@ -61,15 +61,15 @@ static void comm_ws_send(ws_t *ws, hk_ep_t *ep)
         char *name = hk_ep_get_name(ep);
         char *value = hk_ep_get_value(ep);
 	int size = strlen(tile_name) + strlen(name) + strlen(value) + 24;
-        uint64_t t = tstamp_ms();
+        unsigned long long t = tstamp_ms();
 	char str[size];
 
 	/* Send WebSocket event */
         if (hk_tile_nmemb() > 1) {
-                snprintf(str, size, "!%lu,%s.%s=%s", t, tile_name, name, value);
+                snprintf(str, size, "!%llu,%s.%s=%s", t, tile_name, name, value);
         }
         else {
-                snprintf(str, size, "!%lu,%s=%s", t, name, value);
+                snprintf(str, size, "!%llu,%s=%s", t, name, value);
         }
 
 	ws_events_send(ws, str);
